requirements.txt
mcp==1.*
openai>=1.45
python-dotenv
streamlit
pydantic
